Add ability to retrieve the product part number of the paired devices. (#10)
Added a new parameter to the IQDevice class to retrieve the product part number of the paired devices. This will be populated only when using the Connect app with version 5.15 or later.
Removed the outdated copy Comm Watch project and CommWatch.prg. Updated the README to point to the most up-to-date version of the same.